feat(app): neo4j gql entities

+++ apps/mobile_app/lib/app/defaults-providers.dart
@@ -4,7 +4,7 @@ class YahaDefaults {
   final double smallGeoBufferSizeInMeters = 100;
   final double bigGeoBufferSizeInMeters = 1000;
   final double averageSpeedKmh = 4;
-  final double poiImageSearchRadiusInMeters = 50;
+  final int poiImageSearchRadiusInMeters = 50;
 }
 
 final defaultsProvider = Provider<YahaDefaults>(
